backend/requirements.txt
Django>=4.2
djangorestframework>=3.12.4
psycopg2-binary>=2.9.3
django-cors-headers>=3.7.0
python-dotenv
drf_yasg
